Moving Python utility classes to utility header.
