+++ include/lbann/data_readers/data_reader_python.hpp
@@ -29,145 +29,9 @@
 
 #include "data_reader.hpp"
 #ifdef LBANN_HAS_PYTHON
-#include <Python.h>
+#include "lbann/utils/python.hpp"
 
 namespace lbann {
-
-namespace python {
-
-/** @brief Singleton to manage embedded Python session.
- *
- *  This mostly manages the initialization and finalization of the
- *  Python session. It is rarely necessary to interact with the
- *  singleton instance directly.
- *
- *  All static functions are thread-safe.
- */
-class session {
-public:
-
-  /** @brief Start embedded Python session if not already running.
-   *  @details Does nothing if Python has already been started.
-   */
-  static void start_once();
-
-  /** @brief Check if embedded Python session is running. */
-  static bool is_active() noexcept;
-
-  /** @brief Check if a Python error has occurred.
-   *
-   *  Throws an exception if a Python error is detected.
-   *
-   *  @param force_error Whether to force an exception to be thrown.
-   */
-  static void check_error(bool force_error = false);
-
-  /** @brief Get singleton instance.
-   *
-   *  Initializes an embedded Python session the first time it is
-   *  called.
-   */
-  static session& get();
-
-  ~session();
-
-private:
-
-  /** @brief State on main Python thread. */
-  PyThreadState* m_thread_state = nullptr;
-
-  // Lifetime functions
-  session();
-  session(const session&) = delete;
-  session& operator=(const session&) = delete;
-
-};
-
-/** @brief RAII wrapper for Python GIL.
- *
- *  The Python interpreter is not thread-safe, so it uses the "global
- *  interpreter lock" to ensure only one thread is executing at a
- *  time. Multithreading is achieved by periodically transferring
- *  control of the GIL between threads. This makes it hard to get
- *  meaningful speedups from simple multithreading. Certain
- *  operations, e.g. I/O and numerical kernels in NumPy, can be
- *  efficiently parallelized because they yield control of the GIL
- *  while working.
- */
-class global_interpreter_lock {
-public:
-  global_interpreter_lock();
-  ~global_interpreter_lock();
-private:
-  global_interpreter_lock(const global_interpreter_lock&) = delete;
-  global_interpreter_lock& operator=(const global_interpreter_lock&) = delete;
-  PyGILState_STATE m_gil_state;
-};
-
-/** @brief Wrapper around a Python object pointer.
- *
- *  Manages the reference count for a @c PyObject pointer and is
- *  implicitly convertible to the pointer. This is especially
- *  convenient for interacting with Python C API functions that @a
- *  borrow references to their arguments and return @a new references
- *  (this is the most common kind).
- *
- *  Handling reference counts is a tricky part of the Python C API. Be
- *  careful when using functions that @a steal references to their
- *  arguments or return @a borrowed references. See
- *
- *    https://docs.python.org/3.7/c-api/intro.html#reference-counts
- *
- *  for an explanation of reference counts.
- */
-class object {
-public:
-
-  /** @brief Take ownership of a Python object pointer.
-   *  @details @a Steals the reference.
-   */
-  object(PyObject* ptr);
-
-  /** @brief Create a Python string. */
-  object(const std::string& val);
-  /** @brief Create a Python integer. */
-  object(long val);
-  /** @brief Create a Python floating point number. */
-  object(double val);
-
-  object() {}
-  /** @details @a Borrows the reference. */
-  object(const object& other);
-  /** @details @a Borrows the reference. */
-  object& operator=(const object& other);
-  /** @details @a Steals the reference. */
-  object(object&& other) noexcept;
-  /** @details @a Steals the reference. */
-  object& operator=(object&& other);
-  ~object();
-
-  /** @returns @a Borrowed reference. */
-  inline PyObject* get() noexcept                  { return m_ptr; }
-  /** @returns @a Borrowed reference. */
-  inline const PyObject* get() const noexcept      { return m_ptr; }
-  /** @returns @a Borrowed reference. */
-  inline operator PyObject*() noexcept             { return get(); }
-  /** @returns @a Borrowed reference. */
-  inline operator const PyObject*() const noexcept { return get(); }
-
-  /** @brief Release ownership of Python object pointer.
-   *  @returns @a New reference.
-   */
-  PyObject* release() noexcept;
-
-private:
-
-  /** Python object pointer. */
-  PyObject* m_ptr = nullptr;
-
-};
-
-} // namespace python
 
 class python_reader : public generic_data_reader {
 public:
